+++ ricaun.Revit.DB.Shape/Revit/Commands/TessellatedShapeSphereCommand.cs
@@ -33,11 +33,52 @@ namespace ricaun.Revit.DB.Shape.Revit.Commands
 
                 var mesh = tessellatedShape.OfType<Mesh>().FirstOrDefault();
 
-                document.CreateDirectShape(mesh);
+                if (mesh != null)
+                {
+                    document.CreateDirectShape(mesh);
+
+                    document.CreateDirectShape(
+                        TessellatedShapeCreator.CreateMesh(mesh.GetTriangleVertices().ToArray()).GetGeometricalObjects()
+                    ).Location.Move(2 * XYZ.BasisX);
+                }
+
+                document.Regenerate();
 
                 document.CreateDirectShape(
-                    TessellatedShapeCreator.CreateMesh(mesh.GetTriangleVertices().ToArray()).GetGeometricalObjects()
-                ).Location.Move(2 * XYZ.BasisX);
+                    TessellatedShapeCreator.CreateMesh(shape.Clone().GetTriangleVertices(0.3).ToArray(),
+                    materialIds: new ElementId[] { MaterialUtils.CreateMaterial(document, Colors.Index.Color_131).Id }).GetGeometricalObjects()
+                ).Location.Move(4 * XYZ.BasisX);
+
+                shape = shape.Clone();
+
+                //for (double i = 0; i <= 1; i += 0.1)
+                //{
+                //    var vertices = shape.GetVertices(i).ToArray();
+                //    var indices = shape.GetIndices(i).ToArray();
+                //    System.Console.WriteLine($"[{i:0.00}] {vertices.Length} {indices.Length}");
+                //}
+
+                var count = 0;
+                for (double i = 0; i <= 1; i += 0.1)
+                {
+                    var vertices = shape.GetVertices(i).ToArray();
+                    var indices = shape.GetIndices(i).ToArray();
+                    System.Console.WriteLine($"[{i:0.00}] {vertices.Length} {indices.Length}");
+                    document.CreateDirectShape(
+                        TessellatedShapeCreator.CreateMesh(vertices, indices,
+                        materialIds: new ElementId[] { MaterialUtils.CreateMaterial(document, Colors.Index.Color_11).Id }).GetGeometricalObjects()
+                    ).Location.Move((2 * count++) * XYZ.BasisX + 5 * XYZ.BasisY);
+                }
+
+                count = 0;
+                for (double i = 0; i <= 1; i += 0.1)
+                {
+                    var verticeTriangles = shape.GetTriangleVertices(i).ToArray();
+                    document.CreateDirectShape(
+                        TessellatedShapeCreator.CreateMesh(verticeTriangles,
+                        materialIds: new ElementId[] { MaterialUtils.CreateMaterial(document, Colors.Index.Color_21).Id }).GetGeometricalObjects()
+                    ).Location.Move((2 * count++) * XYZ.BasisX + 10 * XYZ.BasisY);
+                }
 
                 transaction.Commit();
             }